zig: promote f32→f64 for mute_freq, threshold args, RecoEvent.score (#377)
Coordinated [cpp-mirror] change with ham PR #21. Closes #377. Removes the
latent narrow-then-widen patterns that PR #376 pinned via stable_sort.

Three sites:

1. Query.mute_freq (query.zig) — was f32, now f64. Field, ctor param, and
   three feed sites updated. Drops three @floatCast narrowings: the two
   Query.create call sites in the initial cachefo build (glomerator.zig:214,
   :236) and the merged-cluster weighted-average computation in
   getMergedQuery (glomerator.zig:1208 → now 1217). qrow.mut_freq is already
   f64 in args.zig, so the @floatCast was unconditionally narrowing on every
   ingest.

2. Four threshold args (args.zig: hamming_fraction_bound_lo/hi,
   logprob_ratio_threshold, max_logprob_drop) — was f32, now f64. CLI
   parser in bcrham.zig now uses parseFloat(f64, ...); the f32_flags tuple
   renamed to f64_flags. The -inf(f32) sentinel comparison at
   glomerator.zig:347 promoted to -inf(f64).

3. RecoEvent.score (reco_event.zig) — was f32, now f64. setScore no
   longer narrows on assignment. This is the field #375 traced to the 30k
   Zig-vs-C++ partition divergence: two ksets with f64 Viterbi scores within
   ~1e-6 collapsed to identical f32 values. Phase 3 pre-work (audit
   partis/python/ for explicit np.float32 casts on logprob) came back
   clean — no round-trip dependency on the partis side.

The std.sort.block in Result.finalize stays — at f64 precision two ksets can
still produce machine-epsilon-equal scores, and we still want deterministic
tie-breaking across stdlib sort algorithms (C++ introsort vs. Zig pdqsort).
Comment block in result.zig and the cache-cascade comment at
glomerator.zig:1463 rewritten to reflect the post-#377 invariant.

submodule packages/ham → 377-float-to-double tip (ham PR #21), which
contains the matching C++ side type changes.

+++ packages/zig-core/src/ham/glomerator/query.zig
@@ -22,7 +22,7 @@ pub const Query = struct {
     seed_missing: bool,
     only_genes: std.ArrayListUnmanaged([]u8),
     kbounds: KBounds,
-    mute_freq: f32,
+    mute_freq: f64,
     cdr3_length: usize,
     /// Names of the two parent queries (may be empty when not set).
     parents: ?[2][]u8,
@@ -51,7 +51,7 @@ pub const Query = struct {
         seed_missing: bool,
         only_genes: []const []const u8,
         kbounds: KBounds,
-        mute_freq: f32,
+        mute_freq: f64,
         cdr3_length: usize,
         parent1: ?[]const u8,
         parent2: ?[]const u8,
